Sanction and Order
Worked on retrieving quote data from the database and displaying it on the sanction landing page and then being able to edit and save changes that have been made to the quotes. Added the feature to send the quotes to be sanctioned will sends the quotes to the order landing page which also has the same feature as the sanction landing with the only difference being it has the ability to order the quotes.